requirements.txt
Flask
Flask-Cors
Flask-PyMongo
pymongo
python-dotenv
gunicorn